requirements.txt
aiogram
aiohttp
asyncpg
asyncio
loguru
python-dotenv
pydantic
pydantic-settings
redis